requirements.txt
langchain>=0.1.0
langgraph>=0.0.30
langchain-openai>=0.0.5
langchain-experimental>=0.0.47
langsmith>=0.0.87
faiss-cpu>=1.7.4
numpy>=1.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
rich>=13.0.0
typer>=0.9.0

